% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMLR_Feature_Ranking.R
\name{CIMLR_Feature_Ranking}
\alias{CIMLR_Feature_Ranking}
\title{CIMLR Feature Ranking}
\usage{
CIMLR_Feature_Ranking(A, X)
}
\arguments{
\item{A}{an (n x n) similarity matrix by CIMLR}

\item{X}{a set of multi-omic data inputs each of which is an (m x n) data matrix of measurements of cancer patients appended by rows}
}
\value{
a list of 2 elements: pvalues and ranking ordering over the n covariates 
as estimated by the method
}
\description{
perform the CIMLR feature ranking algorithm. This takes as input the combination 
of the original input data appended by rows and the corresponding similarity matrix 
computed by CIMLR
}
\examples{
cimlr = CIMLR(X = GliomasReduced$in_X, c = 3, cores.ratio = 0)
input_data = rbind(GliomasReduced$in_X$point_mutations,GliomasReduced$in_X$copy_numbers, 
                   GliomasReduced$in_X$methylations,GliomasReduced$in_X$expression_values)
CIMLR_Feature_Ranking(A = cimlr$S, X = input_data)

}
